MS-Orchestractor/requirements.txt
fastapi[standard]
pydantic
uvicorn[standard]
httpx
email-validator
python-dotenv
